CRUD encerrada

+++ app/Http/Controllers/AlunoCrudController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Aluno;
+use App\Models\AlunoTurma;
 use Illuminate\Http\Request;
 use App\Http\Requests\AlunoRequest;
+use Illuminate\Support\Facades\DB;
 use Alert;
 
 class AlunoCrudController extends Controller
@@ -51,10 +53,18 @@ class AlunoCrudController extends Controller
 
     public function destroy($id)
     {
-        $aluno = Aluno::findOrfail($id);
+        
+        $aluno_select = AlunoTurma::all()->where('id_aluno',$id);
+        if(count($aluno_select) > 0){
+            return redirect()->back()->with([toast()->error('Esse estudante possui matrícula')]);
+        }
+        else{
+            $aluno = Aluno::findOrfail($id);
+            $aluno->delete(); 
+    
+            return redirect()->route('alunos.show')->with([toast()->info('Cadastro de Aluno excluído com sucesso!')]);
+        }
 
-        $aluno->delete();
-
-        return redirect()->route('alunos.show')->with([toast()->info('Cadastro de Aluno excluído com sucesso!')]);
+        
     }
 }
