+++ app/Http/Controllers/TurmaCrudController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Turma;
+use App\Models\AlunoTurma;
 use App\Models\Escola;
 use App\Http\Requests\TurmaRequest;
 use App\Http\Resources\TurmaResource;
@@ -61,10 +62,16 @@ class TurmaCrudController extends Controller
 
     public function destroy($id)
     {
-        $turma = Turma::findOrfail($id);
+        $turma_select = AlunoTurma::all()->where('id_turma',$id);
+        if(count($turma_select) > 0){
+            return redirect()->back()->with([toast()->error('Esse registro possui referências')]);
+        }
+        else{
 
+        $turma = Turma::findOrfail($id);
         $turma->delete($id);
 
         return redirect()->route('turmas.show')->with([toast()->info('Cadastro de turma excluído com sucesso!')]);
+        }
     }
 }
